requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
Django==5.0.7
django-picklefield==3.2
django-q==1.3.9
djangorestframework==3.15.2
doppler-sdk==1.3.0
gunicorn==22.0.0
h11==0.14.0
http-exceptions==0.2.10
idna==3.7
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.4
wcwidth==0.2.13
